Copy webservice tests from pipeline-it
Instead of launching the webservice from the pipeline2 start script in
pipeline-assembly, launch it directly via Java. This allows us to run
the tests within the webservice project itself.

Because both ways of testing are useful and complementary, leave a copy
of the tests in pipeline-it. Ideally the duplicated test code should be
shared somehow.