xproc-engine-daisy-pipeline: make all Pipeline dependencies "provided"
... for the purpose of choosing a version of the Pipeline at runtime,
which is desired because we don't want to release a new
xproc-engine-daisy-pipeline for each new version of framework-bom.

Specifying the Pipeline version is done explicitly by including a
certain version of calabash-adapter on the test classpath. There is no
default version, so if calabash-adapter is not explicitly added to the
classpath, xproc-engine-daisy-pipeline will not work. framework-core,
common-utils and xproc-api are dependencies of calabash-adapter so don't
have to be specified too.

Without this change, by simply importing the right framework-bom in the
dependencyManagement section we may have ended up with the desired
Pipeline artifacts on the test classpath too. This would have been a
more convenient solution, however it would have been less transparent.

See also discussion at https://github.com/daisy/pipeline-build-utils/pull/33