Support environment variables for all supported system properties
in a uniform way.

A utility function was added to read from both System.properties and
System.env. If a property starts with "org.daisy.pipeline", the
function first looks for an environment variable and falls back to a
system property. Otherwise it looks only for a system property.

For example, if the key is "org.daisy.pipeline.ws.host", an
environment variable "PIPELINE2_WS_HOST" will have precedence over the
system property.