Skip persistence-derby tests on Travis
They work fine on my own machine, also with (Oracle) Java 11, but for
some reason they fail on Travis. Possibly it has to do with Travis
using OpenJDK. (Even when you specify "jdk: oraclejdk11", a version of
OpenJDK from java.net is downloaded and installed.)

Note that it is not possible anymore to select a JDK < 9.