Make sure URL objects are always encoded
Unlike the URI class, the URL class does not impose whether fields are
encoded or not. This can result in some confusion, as one can never be
sure about the encoding.

Until now the tendency was to not encode URL objects. However the
problem was that it was not done consistently. And secondly, it
actually makes more sense to encode them.

Also don't make use of URLDecoder anymore because it is only meant to
be used for HTML form encoding (which is not the same as RFC2396).