Improve logging-appender
Replace IgnoreSiftAppender with a more specialized appender named
"JobLogFileAppender" that already knows how to sift job messages and
where to store the log files of each job. This simplifies the
configuration while preserving the possibility to configure the
formatting of messages.

Also give EventBusAppender a more appropriate name:
"JobProgressAppender".

Remove an unneeded dependency.