Implement poor man's version of Saxon's reflexive extension function mechanism
which is only available in the PE and EE versions. See
https://www.saxonica.com/documentation/index.html#!extensibility/functions.

This mechanism has the benefit of being more portable and requiring
less boilerplate code. It'll make it easier to add new XSLT-Java
interop and we can use it to simplify the code base. We can also use
for the SaveAsDAISY to Pipeline 2 port.

See also https://github.com/daisy/pipeline-mod-braille/issues/84.