Make Job extend Runnable
The Job.run(XProcEngine) is removed and replaces by Job.run(). This
means the engine is already embedded in the job object, so one thing
less to care about by the user of the Job API.

Note that an alternative solution could have been to remove the method
completely from the API, as starting the job is already handled by the
JobManager that creates it. However the idea is to also provide a way
to create jobs without a JobManager, and in that case it makes sense
to let the caller deal with creating the thread and starting the job
in it.