Eliminate JobContext interface
The methods were added to Job instead:

- getNiceName()
- getScript()
- getMonitor()
- getLogFile()
- getResults()
- getBatchId()
- getClient()

Also removed AbstractJobContext from the JobStorage
API. DefaultJobManager is now responsible for creating the AbstractJob
and sending it to the JobStorage (which might return a new
AbstractJob).

DefaultJobManager is now also responsible for broadcasting the
initial (idle) status.

JobContextFactory was removed from the public package and integrated
in DefaultJobBuilder. A part of the code of AbstractJobContext was
moved to DefaultJobBuilder as well. Another part was moved to
AbstractJob.